+++ game/content/ghplots/lancedev_objectives.py
@@ -22,6 +22,59 @@ BAME_LANCEMATE = "BAME_LANCEMATE"
 # Personal Lancedev Objectives
 BAMOP_DISCOVER_BIOTECHNOLOGY = "BAMOP_DiscoverBiotechnology"
 BAMO_BETRAYAL = "BAMO_BETRAYAL"
+BAMO_PRACTICE_DUEL = "BAMO_PRACTICE_DUEL"
+
+
+class DDBAMO_PracticeDuel(Plot):
+    LABEL = BAMO_PRACTICE_DUEL
+    active = True
+    scope = "LOCALE"
+
+    def custom_init(self, nart):
+        myscene = self.elements["LOCALE"]
+
+        self.register_element("ROOM", pbge.randmaps.rooms.FuzzyRoom(15, 15, anchor=pbge.randmaps.anchors.middle),
+                              dident="LOCALE")
+
+        team2 = self.register_element("_eteam", teams.Team(enemies=(myscene.player_team,)), dident="ROOM")
+
+        mynpc = self.elements[BAME_LANCEMATE]
+        self.locked_elements.add(BAME_LANCEMATE)
+        self.party_member = mynpc in nart.camp.party
+        if self.party_member:
+            plotutility.AutoLeaver(mynpc)(nart.camp)
+        plotutility.CharacterMover(nart.camp, self, mynpc, myscene, team2, allow_death=False)
+
+        self.obj = adventureseed.MissionObjective("Defeat {}".format(mynpc), missionbuilder.MAIN_OBJECTIVE_VALUE * 2)
+        self.adv.objectives.append(self.obj)
+
+        self.intro_ready = True
+
+        return True
+
+    def _eteam_ACTIVATETEAM(self, camp):
+        if self.intro_ready:
+            npc = self.elements[BAME_LANCEMATE]
+            ghdialogue.start_conversation(camp, camp.pc, npc, cue=ghdialogue.ATTACK_STARTER)
+            self.intro_ready = False
+
+    def BAME_LANCEMATE_offers(self, camp):
+        mylist = list()
+        mylist.append(Offer("[DUEL_GREETING]",
+                            context=ContextTag([context.CHALLENGE, ])))
+        return mylist
+
+    def t_ENDCOMBAT(self, camp):
+        myteam = self.elements["_eteam"]
+        if len(myteam.get_members_in_play(camp)) < 1:
+            self.obj.win(camp, 100)
+        else:
+            self.obj.failed = True
+        if self.party_member:
+            self.elements[BAME_LANCEMATE].restore_all()
+            plotutility.AutoJoiner(self.elements[BAME_LANCEMATE])(camp)
+        # for pc in camp.party:
+        #    pc.restore_all()
 
 
 class LMBetrayalFight(Plot):