+++ history.md
@@ -1,8 +1,9 @@
+* Fixed bug with random MissionText
 * Scenario builder automatically adds QOL reporter to cities
 * Added message log
 
 v0.976 April 10, 2024
-* Fixed a bug in the containers unit!
+* Fixed a bug in the containers unit
 * Merit badges may have contra badges which get removed if badge earned
 * Removed unusued and not working android import from pbge; if android support added later, will need to rewrite anyhow
 * Added config troubleshooting option to disable scaling altogether